adjusted directory structure

+++ torchcor/ionic/MitchellSchaeffer.py
@@ -1,10 +1,10 @@
 import torch
-from ionic.cellml import mitchell_schaeffer_2003
-from ionic.base import BaseCellModel, BaseCellModelRL
+from torchcor.ionic.cellml import mitchell_schaeffer_2003
+from torchcor.ionic.base import BaseCellModel, BaseCellModelRL
 
 
 class MitchellSchaeffer(BaseCellModel):
-    def __init__(self, device, dtype=torch.float64):
+    def __init__(self, device, dtype=torch.float32):
         super().__init__(mitchell_schaeffer_2003, device, dtype)
         self.name = "MitchellSchaeffer"
